Containerise algorithms service
We update the algorithms service dockerfile to leverage caching for
faster build times, use a non-privileged user for running to improve
security, and add an entrypoint.sh script to start the service.

Additionally, we modify the docker volumes to use local mounts.
This means that developers can easily inspect the contents of the
files and even delete from the worktree. Naturally these files
were also added to .gitignore.